+++ trax/CellBuilder.h
@@ -9,7 +9,7 @@ class Range;
 namespace CellBuilder
 {
 void isoband_linear(JointMerger& joints, const Cell& c, const Range& range);
-void isoline_linear(JointMerger& joints, const Cell& c, double limit);
+void isoline_linear(JointMerger& joints, const Cell& c, float limit);
 void isoband_midpoint(JointMerger& joints, const Cell& c, const Range& range);
 
 }  // namespace CellBuilder
