+++ trax/GridPoint.h
@@ -5,9 +5,9 @@ namespace Trax
 struct GridPoint
 {
   GridPoint() = default;
-  GridPoint(float x, float y, float z) : x(x), y(y), z(z){};
-  float x = 0;
-  float y = 0;
+  GridPoint(double x, double y, float z) : x(x), y(y), z(z){};
+  double x = 0;
+  double y = 0;
   float z = 0;
 };
